updated layout to container/form/index
