+++ src/components/CreateTeam/container.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { PropTypes as PT } from 'prop-types';
 import { createTeam } from '../../services/teamService';
+import { getAllClients } from '../../services/clientService';
 import swal from 'sweetalert2';
 import { Toast } from '../../config/Notifications';
 
@@ -15,6 +16,40 @@ export default Wrapped =>
 
     state = {
       teamSubmitted: false,
+      clients: [],
+    }
+
+    componentDidMount() {
+
+      getAllClients()
+        .then(response => {
+          const clients = response.data;
+          const formattedClients = clients.reduce((acc, client) => {
+            acc.push({
+              value: client.clientId,
+              displayValue: client.clientName,
+            });
+            return acc;
+          }, []);
+          formattedClients.unshift({
+            value: -1,
+            displayValue: 'Please select a client',
+          });
+          this.setState({ clients: formattedClients });
+        })
+        .catch(error =>
+          swal(
+            'Error',
+            `Could not retreive clients: ${error.message}`,
+            'error',
+          ),
+        );
+    }
+
+    shouldComponentUpdate(nextProps, nextState) {
+      const teamSubmittedhasChanged = nextState.teamSubmitted !== this.state.teamSubmitted;
+      const clientListhasChanged = nextState.clients.length !== this.state.clients.length;
+      return teamSubmittedhasChanged || clientListhasChanged;
     }
 
     submitRequest = data => {
@@ -45,11 +80,16 @@ export default Wrapped =>
     };
 
     render() {
+
+      const { clients, teamSubmitted } = this.state;
+      const { history: { replace } } = this.props;
+
       return (
         <Wrapped 
-          navigateTo={this.props.history.replace} 
+          clients={clients}
+          teamSubmitted={teamSubmitted}
+          navigateTo={replace} 
           submitRequest={this.submitRequest} 
-          teamSubmitted={this.state.teamSubmitted}
           resetTeamSubmitted={this.resetTeamSubmitted}
         />
       );
