+++ src/components/CreateTeam/index.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { PropTypes as PT } from 'prop-types';
 import container from './container';
-import CreateTeamForm from './CreateTeamForm';
+import CreateTeamForm from './form';
 import { Button } from '../common';
 import { CornerButton } from '../common_styled';
 
-export const CreateTeam = ({ submitRequest, navigateTo, teamSubmitted, resetTeamSubmitted }) => {
+export const CreateTeam = ({ clients, submitRequest, navigateTo, teamSubmitted, resetTeamSubmitted }) => {
   return (
     <div>
       <CornerButton>
@@ -15,16 +15,21 @@ export const CreateTeam = ({ submitRequest, navigateTo, teamSubmitted, resetTeam
         />
       </CornerButton>
       <h2>Create Team</h2>
-      <CreateTeamForm onSuccess={submitRequest} teamSubmitted={teamSubmitted} resetTeamSubmitted={resetTeamSubmitted} />
+      <CreateTeamForm clients={clients} handleFormSubmit={submitRequest} teamSubmitted={teamSubmitted} resetTeamSubmitted={resetTeamSubmitted} />
     </div>
   );
 };
 
 CreateTeam.propTypes = {
+  clients: PT.array,
   submitRequest: PT.func.isRequired,
   navigateTo: PT.func.isRequired,
   teamSubmitted: PT.bool.isRequired,
   resetTeamSubmitted: PT.func.isRequired,
+};
+
+CreateTeam.defaultProps = {
+  clients: [],
 };
 
 export default container(CreateTeam);
